Main 파일 수정
Video 수정

+++ hantalk/src/main/java/com/example/hantalk/repository/VideoRepository.java
@@ -3,5 +3,5 @@ package com.example.hantalk.repository;
 import com.example.hantalk.entity.Video;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface VideoRepository extends JpaRepository<Video, Long> {
+public interface VideoRepository extends JpaRepository<Video, Integer> {
 }
